mpc_controller/robot-config-project/requirements.txt
numpy
pyyaml
robotics-toolbox-python
urdfdom-py
pytest